backend/requirements.txt
numpy
pandas
scikit-learn
tensorflow
keras
pytz
joblib
google-generative-ai
